+++ server/suspensions_test.go
@@ -11,14 +11,14 @@ func TestSuspender(t *testing.T) {
 	s := newSuspender()
 
 	s.suspend("foo", 5)
-	assert.Equal(s.Suspended("foo"), int64(5))
+	assert.Equal(s.Suspended("foo"), 5)
 	s.suspend("foo", 5)
-	assert.Equal(s.Suspended("foo"), int64(10))
+	assert.Equal(s.Suspended("foo"), 10)
 	s.count = 11
-	assert.Equal(s.Suspended("foo"), int64(0))
+	assert.Equal(s.Suspended("foo"), 0)
 	_, ok := s.list["foo"]
 	assert.False(ok)
 
 	s.signalRefresh()
-	assert.Equal(s.count, int64(12))
+	assert.Equal(s.count, 12)
 }
